requirements.txt
fastapi==0.73.0 
uvicorn[standard]==0.17.5 
psycopg2==2.9.3
pandas==1.4.1
PyYAML==6.0 